Update Email formatting
- Added CSS for update email
