+++ .gitignore
@@ -41,4 +41,4 @@ next-env.d.ts
 
 server.js
 
-new-email
+email-testing
