+++ pages/api/update-notification.js
@@ -1,23 +1,10 @@
+import ReactDOMServer from "react-dom/server";
 import NodeMailer from "nodemailer";
+import UpdateEmail from "@/components/update-email";
 import { fetchSingleRequest } from "@/lib/fetchRequests";
 
 export default async function handler(req, res) {
   const { id, token } = req.body; // TODO: refactor this to only need ID. This will involve changing the values below to grab title and status for the email
-
-  function statusColor(status) {
-    switch (status) {
-      case "New":
-        return "#ff52d9";
-      case "In Progress":
-        return "#7480ff";
-      case "Pending":
-        return "#ffbe00";
-      case "Complete":
-        return "#00a96e";
-      default:
-        return "";
-    } // The colors here are subject to change depending on if the theme changes
-  }
 
   const transporter = NodeMailer.createTransport({
     service: "gmail",
@@ -31,6 +18,16 @@ export default async function handler(req, res) {
     // Grab the email from the backend so email is not exposed in json body
     const emailData = await fetchSingleRequest(id, token);
 
+    // Generate email HTML
+    const htmlContent = ReactDOMServer.renderToStaticMarkup(
+      <UpdateEmail
+        title={emailData[0].request_title}
+        status={emailData[0].request_status}
+        note={emailData[0].request_note}
+        image={""}
+      />
+    );
+
     // Verify transporter connection before sending notification
     await transporter.verify();
 
@@ -38,15 +35,7 @@ export default async function handler(req, res) {
       from: "PlexRequest Notification <bm.contact623@gmail.com>",
       to: emailData[0].request_requestor,
       subject: `${emailData[0].request_title} Update!`,
-      html: `
-      <h2>Here is an update on your request for: ${emailData[0].request_title}</h2>
-      <h2>Status: <span style="border: 3px solid ${statusColor(
-        emailData[0].request_status
-      )}; border-radius: 5px;">
-      ${emailData[0].request_status}
-      </span></h2>
-      ${!emailData[0].request_note ? "" : `<h2>Note: ${emailData[0].request_note}</h2>` }
-      `, // TODO: add CSS to make this look nicer
+      html: htmlContent,
     };
 
     // Send notification
